requirements.txt
streamlit
streamlit-authenticator
firebase-admin
pyrebase4
requests
PyYAML